+++ lending/RELEASES.md
@@ -79,3 +79,13 @@ Based on:
 - [typescript v3.3.1] lending
 ### Releases
 - [NPM v3.3.1] https://www.npmjs.com/package/@codat/lending/v/3.3.1 - lending
+
+## 2023-09-28 00:04:49
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.91.2 (2.131.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.3.2] lending
+### Releases
+- [NPM v3.3.2] https://www.npmjs.com/package/@codat/lending/v/3.3.2 - lending